Proper language for data-record definition
The data records draft talked about IOAM
information being encoded as TLVs - which is not
really the case, given that IOAM records are of
fixed size. The only variable is how many are
embedded (which is a lenght variable) and which
are embedded (which is determined by a bit field).

This patch adjusts the language in the data draft
accordingly.

See also https://github.com/inband-oam/ietf/issues/35

Signed-off-by: Frank Brockners <fbrockne@cisco.com>