Processing undefined IOAM-Trace-Type bits
Define behavior when receiving a packet with any of the undefined
IOAM-Trace-Type bits set. This will allow a network to run with a transit
node running this revision, and an encapsulating node running a future
revision that is compatible but supports new definitions for some of the
currently undefined IOAM-Trace-Type bits. In this case the transit node
will not know how to add node data fields for the IOAM-Trace-Type bits
defined in the future revision.

This is a proposed resolution for issue #63.