Replace Incremental Trace Max Length with Remaining Length
For in-situ OAM Incremental Trace, the current draft specifies a Max Length
field that limits the size of the node data list. While this achieves the
goal of limiting the impact of in-situ OAM on data packets, it requires a
comparison between 3 or 4 values plus a constant in order to determine
whether an IOAM transit node is allowed to add its node data to the list:

    Max Length - (Trace Shim Header Length - 1) >=
    NodeLen + IOAM-Trace-Type Bit 7 ? Opaque State Snapshot Length : 0

This can be simplified by replacing Max Length with a Remaining Length
field:

    Remaining Length >=
    NodeLen + IOAM-Trace-Type Bit 7 ? Opaque State Snapshot Length : 0

Note: Proposed terminology is "RemainingLen" rather than "Octets-Left"
terminology used in Pre-Allocated Trace Option, since "Octets-Left" is
specified in multiples of 4-octets. That seems confusing. Preferrably the
Pre-Allocated Trace Option and the Incremental Trace Option should converge
to the same terminology.