update the definition of queue depth' units
There is a problem with the 08 definition of queue depth, as it allows too much variation in the reporting of the queue length. This may cause the prevention of interop across implementations.
Suggesting to align the implementations on a standard memory unit that will enable interop, across product generations and across vendors.