Revert O-bit value "1" to mean overflowed
An "editorial" change modified the text about setting the O-bit, saying
that the bit should be set to "0". This is non-intuitive, and leaves it
unclear when the bit would be set to "1".

This commit changes the text to set the bit to "1" when there are
insufficient octets to record node data.