Minor changes
Quick first pass, some minor changes applied. I kept the IPv6
encapsulation section as is but, for now, it is not included in the
kernel implementation (only the direct insertion is supported). However,
this is just a matter of time so I guess it's OK to keep it here until
that. Same remark applies for both the queue depth and buffer occupancy
data fields.

Signed-off-by: Justin Iurman <justin.iurman@uliege.be>