Fix validator algorithm (IOAM integrity)
There is a problem with the current method defined for the validator.
Indeed, for the asymmetric approach, the validator only receives public
keys from the nodes. So, basically, it is incorrect to say that it can
reproduce the signature chain from encap to decap (in that order) and
compare it to the signature field. This method would only be valid in a
scenario using the symmetric approach. The only possibility for it to be
valid for asymmetric as well would be that the validator receives
private keys too (which is definitely not what we want). As a
consequence, that very specific part was rewritten to fix this.

Note: this method automatically solves the corner case when there is a
Trace Option-Type that requires the Opaque State Snapshot.