-12: keywords, option names, "In Situ" alignment
- improved keywords
- option names moved from "IOAM XXX Integrity Protected Option-Type" to
  "IOAM Integrity Protected XXX Option-Type"
- be consistent with other RFCs by using "In Situ" (not "in-situ", for
  example), by expanding OAM in the title, etc

Note: the renaming of options was needed to highlight the main
"category" of these options, as being "Integrity Protected" ones
(instead of having it at the end, which is confusing). Plus, it will
ease to align draft-ioam-integrity-yang (draft-to-be) with RFC9617. IANA
has been notified and they're OK with the change. Amanda's reply quoted:

"In this case, because this is just a minor wording change, you can just
send us a note when the names have been updated in the document."